+++ packages/superset-ui-chart-controls/src/operators/rollingWindowOperator.ts
