+++ packages/superset-ui-chart-controls/src/utils/index.ts
@@ -20,6 +20,5 @@ export * from './selectOptions';
 export * from './D3Formatting';
 export * from './expandControlConfig';
 export * from './getColorFormatters';
-export * from './operators';
 export { default as mainMetric } from './mainMetric';
 export { default as columnChoices } from './columnChoices';
